007-03 gemini_asst_revised/requirements.txt
discord.py>=2.3.0
google-generativeai>=0.3.0
notion-client>=2.2.0
python-dotenv>=1.0.0